Remove dtc::string and use std::string.  NFC.
dtc::string was designed to keep a reference to a contiguous sequence of bytes stored in an input buffer, eliminating the need to copy (until the final output).  dtc::string was very lightweight (a pointer and a length) and so was cheap to pass around by value.  In contrast, std::string owns the text and is far heavier.

These changes are needed to fully support the horrible /input/ semantics that recent versions of the dts format have adopted, as we no longer have a guarantee that ranges that are currently stored as strings are going to be contiguous.  The alternative would be to run the input through a preprocessor prior to parsing, but that would likely be no more efficient.