Move the inclusion handling lower down.
The new `text_input_buffer` class manages a stack of input buffers.  When moving to the next token, it will check the input stream for `/include/` tokens and handle them, pushing a new buffer onto the stack.

This also reworks the error reporting somewhat, so that we now report the file that triggered the error.

Fixes #4