Implement the already declared but never defined device_tree::assign_phandles()
Have it assign a phandle to the current node if it has any labels, then walk its
children regardless and recursively assign phandles. Using this method will give
us a situation where labels are basically assigned in order that we come across
them in the source tree, but this fact should not be relied upon in consuming
implementations.