Use device_tree::assign_phandles() if we're writing symbols (dtc -@)
This matches behavior of GPL dtc(1). While I don't agree with it, Linux and
libfdt are both going to be assuming that all labelled nodes in an FDT blob
will have phandles assigned to them if we've generated a symbol table.

We still pass the next to-be-assigned phandle to
device_tree::resolve_cross_references(), but it should remain effectively unused
there given that we've already assigned all phandles. It seems to be the
cleanest method of implementing this feature so that things still work as
expected in both the symbol case and the non-symbol case