Add generate_root function for generating a proper initial root from a node
This function is aware of whether it's in a plugin or not so that it may do the
right thing. If it's not in a plugin and it's given a potential root that isn't
named "", it bails out and we throw a parse error. If we're in a plugin, we
wrap the root in a fragment, attach it to a root node, then return the new root
node.

This is only used for the initial roots of a tree, where resolution to names
already defined is not done because there haven't really been any names to
encounter. Later roots will be taken care of in future work.