Handle absolute paths properly for included files
This fixes issue #25, reported by manu@. u-boot-pine64 uses absolute paths to
include its data, which get broken when we try to get an input buffer for them
as it prepends the directory.

Force it to use absolute paths as-is, and don't fallback to include paths if
given an absolute path. If an included file is accidentally found in an include
path, it may not necessarily do the right thing and cause more harm than good.