Update, still broken
Notable changes:
- Renamed to /omit-if-no-ref/, per GPL DTC
- Rewrote approach to align more closely with David's vision

node::delete_child turned into node::delete_children_if to allow the GC
process to just use a simple predicate to determine that one of the children
needs to go away.

This still segfaults, later on in the checking phase. More care needs to be
taken when removing children from the tree, as there are other references that
also need to be cleaned up in, for example, node_names and node_paths. The
order in which garbage collection occurs also needs more thought put into it,
since symbols have already been written. We should skip symboled nodes by
coincidence at the moment because we traverse /__symbols__ in our initial visit,
but this should really occur before symbols are written and knowingly either
emit or omit nodes with symbols.

Some discussion was occurring in GPL dtc on whether or not to ignore the flag
on nodes with symbols. It looks like they might settle on another program
argument to keep them, because the entire point of /omit-if-no-ref/ in the first
place was for use on pinctrl nodes that will need a label.