Mark the device_tree as needing garbage collection during parsing
Garbage collection is a quite expensive process, involving two tree walks at a
minimum at the moment.  This could be optimized to not do the second walk if the
first walk didn't encounter any nodes needing garbage collection, we can do
better than that by simply not bothering with it if we didn't encounter any
/omit-if-no-ref/ nodes (the common case).

The node ctor and node::parse needed to learn about the device_tree in order to
indicate that an /omit-if-no-ref/ was encountered.  Most other approaches would
still require either additional tree walks or for the omit-if-no-ref property to
be tracked yet again up the node hierarchy until the root of the section of the
tree it occurs in.  The author is unconvinced that either of these approaches
are any cleaner.