Fix ordering of /delete-property/ and /delete-node/ processing
These directives are loosely defined to delete previous instances of properties
and nodes, but nothing is explicitly said about what happens if the property or
node is reincarnated after one of these directives appears.

Currently, /delete-property/ and /delete-node/ are processed *after* all of the
'other' properties and children have been merged in. This doesn't match GPL dtc,
which takes the lax interpretation and allows a deleted node to be redefined
immediately after that /delete-node/, so make ourselves compatible.