dtc: assume proper -I/-O formats if only one is specified
This matches the behavior of GPL dtc; for instance, if -I dts is specified,
then -O dtb is assumed. -I dtb assumes -O dts, and all of this works in
reverse as well if just -O is specified.

We still default to -I dts -O dtb if both are omitted or -I dts if -O asm is
specified.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>