Validate that parsed integers fit in cells of a given size
Currently we silently truncate integers, but GPL dtc will give an error
if the integer does not fit, so do the same.

Note that, because we operate on unsigned long long values and lose
track of the sign, we cannot distinguish, say, -1 from 2^64-1. Thus we
are forced to fail to diagnose values in [2^64-2^(n-1), 2^64) as that is
the range negative inputs will be mapped to, whereas GNU dtc is able to
catch those cases.