Validate that integers in the input are not too big for unsigned long long
If the input is too big, strtoull will clamp to ULLONG_MAX and set errno
to ERANGE, so catch this error and fail to parse in that case. The error
reported is not great, but it's better than no error, and the current
code structure makes it awkward to give more detailed info.