Validate integers fit in cells (#65)
Currently we silently truncate integers, but GPL dtc will give an error
if the integer does not fit, so do the same.

Note that, because we operate on unsigned long long values and lose
track of the sign, we cannot distinguish, say, -1 from 2^64-1. Thus we
are forced to fail to diagnose values in [2^64-2^(n-1), 2^64) as that is
the range negative inputs will be mapped to, whereas GNU dtc is able to
catch those cases.

If the input is too big, strtoull will clamp to ULLONG_MAX and set errno
to ERANGE, so catch this error and fail to parse in that case. The error
reported is not great, but it's better than no error, and the current
code structure makes it awkward to give more detailed info.